Moved some code out of the static ifs.
Those static ifs were for code that needed the fixes in #9393 to work.
The code that I moved out of them didn't really need that fix, since it
relied on the third overload of ifThrown(the delegate version without
templating the caught exception).

Also, I changed the comment on those static ifs. I thought it will be
pulled in time for the dmd v2.062 release, but it wasn't.